Applies Nico's comments
